+++ packages/stream_chat_flutter/CHANGELOG.md
@@ -7,12 +7,12 @@
 
 ✅ Added
 
-- Introduced `StreamMessageInput.useNativeAttachmentPicker` for system media picker support.
+- Introduced `StreamMessageInput.useSystemAttachmentPicker` for system media picker support.
 
 🔄 Changed
 
 - Updated the message list view to prevent pinning messages that have restricted visibility.
-- Deprecated `StreamMessageInput.useNativeAttachmentPickerOnMobile` in favor of `StreamMessageInput.useNativeAttachmentPicker`.
+- Deprecated `StreamMessageInput.useNativeAttachmentPickerOnMobile` in favor of `StreamMessageInput.useSystemAttachmentPicker`.
 
 ## 9.4.0